Register Page, Login, Contact Creation
I’ve created the Registration Page, the Login Page ( Created a new jsx file, don’t know how to link it to the App.js page to make it show, but it shouldn’t matter as we are separating the pages. ) and the Contact Form ( not sure why I cannot figure out how to move the form to the left of the page as I make it work for the Registration Form. I’m probably just too tired ). I’ve also created a Mock page for components trials.